requirements.txt
django
djangorestframework==3.14.0
mysqlclient==2.2.1
mysql-connector-python==8.3.0